refactor: move collection en/decodable impls to separate file
